server/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyascii==0.3.2
autocorrect==2.6.1
blinker==1.7.0
blis==0.7.11
cachelib==0.13.0
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
contractions==0.1.73
cymem==2.0.8
Deprecated==1.2.14
dnspython==2.6.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
Flask==3.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
limits==3.11.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mongoengine==0.28.2
msgspec==0.18.6
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
packaging==24.0
preshed==3.0.9
pyahocorasick==2.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pyenchant==3.2.2
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.6.3
pytz==2024.1
regex==2024.4.28
requests==2.32.3
rich==13.7.1
scikit-learn==1.4.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
textsearch==0.0.24
thinc==8.2.4
threadpoolctl==3.5.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.1
wrapt==1.16.0
